Simplify checks by extracting common functions
Checker defines the general interface for Repoman
checks, leaving each checking class to only
implement its own, unique functionality.
