+++ lib/repoman.rb
@@ -8,5 +8,6 @@ require 'open-uri'
 module Repoman
 end
 
+require_relative './repoman/checker'
 require_relative './repoman/file_fetcher'
 require_relative './repoman/license_checker'
